Fix FCFS scheduler and add the double instead of unsigned long long

+++ Core/Process/priority_process.cpp
@@ -1,7 +1,6 @@
 #include "priority_process.h"
 
-PriorityProcess::PriorityProcess(QString name, long long burstTime, unsigned long long arrivalTime, unsigned long long priority):
-    Process(name, burstTime, arrivalTime), priority{priority}
+PriorityProcess::PriorityProcess(QString name, double burstTime, double arrivalTime, unsigned long long priority) : Process(name, burstTime, arrivalTime), priority{priority}
 {
 }
